Remove thrift dependency from TraceSpec
